+++ src/schema.rs
@@ -2,6 +2,6 @@ table! {
     records (id) {
         id -> Integer,
         timestamp -> Text,
-        voltage -> Integer,
+        voltage -> Float,
     }
 }
